backend/requirements.txt
Django==3.1.2
coverage==5.3
pylint==2.6.0
pylint-django==2.3.0
coveralls==2.1.2
phonenumbers==8.12.12
django-phonenumber-field==5.0.0
django-cors-headers==3.5.0
boto3==1.16.25
psycopg2-binary==2.8.6
django-dotenv==1.4.2
awsebcli==3.19.2